[SW-318] Keep H2O version inside sparklin-water-core.jar and provide utility to query it
