+++ .gitignore
@@ -1,5 +1,6 @@
 # generated resource file containing Spark version for which Sparkling Water was built
 /core/src/main/resources/spark.version
+/core/src/main/resources/h2o.version
 *.class
 
 # Mobile Tools for Java (J2ME)
